Tweak the now.json to pickup the app
Adding secrets to the env mapping

Fix the secret naming

Adding the Readme and tests

Update readme and add wrap dotenv with NODE_ENV to load it only for the
local env

Adding the test for the retriable mailer

Push the scope of the mail clients to RetriableMailer. Limits the
dependency and exposure.

Email client tests

Update the documentation

Fixes and updates

Remove the json tag